frontend/apps/graph-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.14.0
python-dotenv==1.0.0
